Removed old accordion poll implementation, removed unused poll swearwords, added proper error handling, fixed poll end events
