+++ assets/swearwords/swearwords.sample.json
@@ -21,10 +21,9 @@
     },
     "other": {
       "title": "Баян?!",
+      "explanation": "Как ты думаешь, это баян или нет?",
       "vote_in_progress": "Голосование уже запущено",
-      "message_not_chosen": "Выбери сообщение, за которое ты хочешь проголосовать",
-      "bot_vote_attempt": "Бот никогда не постит баяны",
-      "explanation": "Как ты думаешь, это баян или нет?"
+      "bot_vote_attempt": "Бот никогда не постит баяны"
     }
   },
   "chat": {
