+++ lib/src/globals/poll.dart
@@ -10,5 +10,7 @@ abstract class Poll {
 
   Duration get duration;
 
+  void endPoll();
+
   void updatePollOptionCount(String option, [int? newOptionResult]);
 }
